Add settings file initialization. * Move logging initialization to utils * Add default settings file
