+++ default_settings.json
@@ -1,0 +1,4 @@
+{
+  "timezone": "US/Eastern",
+  "city": "Toronto"
+}
