+++ main.py
@@ -1,33 +1,11 @@
 # -*- coding: utf-8 -*-
 
-import logging
-
 from thermostat.thermostat import Thermostat
-from thermostat import errors
-
-
-def init_logging():
-    """ Set up logging.
-
-    Logs INFO level or higher to file and DEBUG level or higher to console.
-    """
-    formatter = logging.Formatter('%(asctime)s[%(levelname)s][%(name)s.%(funcName)s] %(message)s')
-    fh = logging.FileHandler('app.log')
-    fh.setLevel(logging.INFO)
-    fh.setFormatter(formatter)
-    ch = logging.StreamHandler()
-    ch.setLevel(logging.DEBUG)
-    ch.setFormatter(formatter)
-
-    logger = logging.getLogger('app')
-    logger.setLevel(logging.DEBUG)
-    logger.addHandler(fh)
-    logger.addHandler(ch)
-    return logger
+from thermostat import errors, utils
 
 
 def main():
-    logger = init_logging()
+    logger = utils.init_logging()
 
     t = Thermostat()
